requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
discord.py==1.7.3
idna==3.3
multidict==6.0.2
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.20.0
pytube==12.1.0
pytube3==9.6.4
requests==2.28.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.10
yarl==1.7.2
